Components now use Restricted instead of isAdmin function
