+++ src/components/navbar/control.tsx
@@ -2,7 +2,7 @@
 
 import React, { ReactNode, useState } from 'react';
 import SegmentedControl from 'components/input/segmented-control';
-import { TabValue } from './content';
+import { TabValue } from '.';
 import { usePathname } from 'next/navigation';
 
 interface NavbarContentProps {
@@ -14,10 +14,9 @@ const NavbarControl = ({ userTab, adminTab }: NavbarContentProps) => {
   const pathname = usePathname();
   const initialTab = pathname?.startsWith('/admin') ? 'admin' : 'user';
   const [tab, setTab] = useState<TabValue>(initialTab);
-  const isAdmin = adminTab !== undefined;
   return (
     <div className='flex flex-col gap-2'>
-      {isAdmin && (
+      {adminTab !== undefined && (
         <SegmentedControl
           color='red'
           defaultValue={initialTab}
