--- src/server/trpc/router/role.ts
@@ -1,11 +1,0 @@
-import { router, publicProcedure } from '../trpc';
-
-export const roleRouter = router({
-  getAll: publicProcedure.query(async ({ ctx }) => {
-    return ctx.prisma.role.findMany({
-      orderBy: {
-        id: 'asc',
-      },
-    });
-  }),
-});
